chore: modificar logica inicial
